@@ -41,6 +41,7 @@
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "babel-preset-stage-2": "^6.18.0",
+    "create-react-class": "^15.5.2",
     "exports-loader": "^0.6.2",
     "file-loader": "^0.10.1",
     "highlight.js": "^9.10.0",
@@ -49,10 +50,10 @@
     "mocha": "^3.2.0",
     "mock-require": "^2.0.1",
     "nightwatch": "^0.9.6",
+    "prop-types": "^15.5.8",
     "raw-loader": "^0.5.1",
-    "react": "^15.4.2",
-    "react-addons-test-utils": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-highlight": "^0.9.0",
     "sinon": "^2.0.0",
     "webpack": "^2.2.1",
